+++ src/DDTrace/Encoders/Json.php
@@ -15,7 +15,7 @@ final class Json implements Encoder
      */
     public function encodeTraces(Tracer $tracer)
     {
-        $json = json_encode($tracer->asArray());
+        $json = json_encode($tracer->getTracesAsArray());
         if (false === $json) {
             self::logDebug('Failed to json-encode trace: ' . json_last_error_msg());
             return '[[]]';